% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{get_country_coordinates}
\alias{get_country_coordinates}
\title{Function that coordinates of country}
\usage{
get_country_coordinates(name)
}
\arguments{
\item{name}{name of the country}
}
\value{
coordinates
}
\description{
Function that coordinates of country
}
