% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{get_country_names}
\alias{get_country_names}
\title{Function that gets all the country names}
\usage{
get_country_names()
}
\value{
json_data_frame
}
\description{
Function that gets all the country names
}
