% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{req_stamen_map}
\alias{req_stamen_map}
\title{Function that interacts with the API}
\arguments{
\item{left}{the left coordinate of map}

\item{bottom}{the bottom coordinate of map}

\item{right}{the right coordinate of map}

\item{top}{the top coordinate of map}

\item{zoom_val}{the zoom}

\item{map_type}{the type of map}
}
\value{
stamenmap
}
\description{
function that interacts with the API
}
